Shader / pipeline hot reloading
Notes:
Pipelines now work very similar to buffers / textures as the renderer has a vector of pipelines and the create_pipeline function returns a index into this array.
This is not the best architecture but vastly simplifies code for the current engine scope.
Shaders have also been modified slightly to allow for shader reloading