Added light abstraction and improved shader hot-reloading
When the new shader of a pipeline fails to compile, rather than throwing a error and terminating the program,
the old pipeline (which was error-free syntactically) will continue to be used, and the user will get warnings
indicates why the new shader failed. This change has been done so as to have a smooth shader-modification process
where the program need not have to be continuously re-opened / terminated in case of shader errors.